workqueue: namespace support
Add support to the kernel workqueue system to allow running a particular
piece of work in a specific set of namespaces. More specifically, the
new queue_work_in() function will associate an nsproxy structure with
the work_struct. Just before calling the work function, the workqueue
thread will switch the current nsproxy to the one associated with the
work_struct. This allows the work function to assume a valid set of
current namespaces just as though it had been called from user space.

Change-Id: Iace39d54e92e3454e2cf788a9bf1c1f68ca0ea86
Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>