dev_ns: bugfix! several refcounting errors fixed
This patch addresses several reference counting and memory corruption
bugs in the device namespace code:

New device namespace would take a reference to the old pid namespace
(the one being copied), not the new one. This resulted in an extra
refcount on the old pid_ns and not enough on the new one. This could
lead to memory corruption.

The dev_namespace structure kept a pointer to all new dev_ns_info
structures created with new_dev_ns_info(), but failed to bump the
reference count. This could lead to memory corruption.

When destroying a dev_namespace structure, the reference to all
dev_ns_info structures kept within the structure must be released.

Change-Id: I11487377d07e086ef19eedd0aadf65d4d1bc802d
Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>