dev_ns: split up dev_ns desc list locking
The static dev_ns_desc list was protected with a spinlock that was also
protecting the individual descriptor object's list of device-specific
dev_ns_info pointers. This has been split into a spinlock to protect the
static list of dev_ns_desc structures (serialize the dev_ns_id values),
and a per-descriptor mutex which is taken when adding to, removing from,
or looping over the list of dev_ns_info structures associated with a
given dev_ns_desc structure, i.e., the list of all device-specific
structures associated with a given device driver.

The second lock needs to be a mutex because there are drivers, such as
the tegra display controller, which need to do possibly blocking work in
the loop_dev_ns callback function.

Note that it is safe to loop through all the dev_ns_info structures in a
given dev_ns_desc structure without holding the dev_ns_desc spinlock
because we assume, as is assumed elswhere, that the function will only
be called when a valid dev_ns_desc object has been registered and
associated with the given dev_ns_id.

Change-Id: Iee59652621721095c1d1493cd6131c5e3644ccae
Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>