LSM: Macroize security hooks
The security hooks are called by accessing the relevant
entry in the security_ops structure. Placing the calls into
a macro does nothing to the performance (it should generate
identical code) and sets the stage for a more sophisticated
calling scheme.

The 2nd parameter to call_int_hook is always 0. It represents
the return value when there is no module supplying the hook.
There is always a hook in the capability module, so this value
is currently unused.

Change-Id: Id0f8ecc2858b12240352b3324c6893070c66c26d
Signed-off-by: Michael Weiss <michael.weiss@aisec.fraunhofer.de>