LSM: Refactor existing LSM stacking
The existing code stacks up to three security modules.
If there is only one module it is the capability module.
If a module is specified the capability code is explicitly
called within the specified module. If Yama stacking is
configured the Yama hooks are called and then the specified
module. The capability check is made twice in the Yama
stacking case.

This patch adds a list for each of the LSM hooks. The capability
hooks that are non-trivial are put on the list first. If Yama
stacking is configured the yama hooks are added to the list.
If a module is configured it gets added next. Each entry in the
list is called in order:

	(capability[ yama][ {selinux|smack|tomoyo|apparmor}])

You only get one of SELinux, Smack, TOMOYO and AppArmor. Addressing
this shortcoming is deferred until this work is complete.

Exit is on first failure (bail on fail) which matches the
existing behavior. The explicit calls to the capability module
are removed where they would duplicate the stacked call. Most
instances of hooks in the capability module simply set a return
value. That is now handled in the "INC" parameter of the
call_int_hook() macro.

The data type of the hook list entries differs, so there has
to be a unique structure for each hook list. That is addressed
in the HOOK() and HOOK_HEAD() macros. It would be simple enough
to expand these in the code, but this is much cleaner.

This is a slightly different (smaller) version of the same patch
posted as 2/3. It works the same.

Change-Id: I6603c5ba4e3e5afe3ba50ad5cf4a2dda6174f361
Signed-off-by: Michael Weiss <michael.weiss@aisec.fraunhofer.de>