trustme-lsm: added path handling+whitelist and priv-drop
This patch adds actual functionality to the trustme-lsm. It introduces a
consolidated handler for all path-based hooks which allows or denies
accesses to paths based on a hard-coded whitelist. It furthermore
prevents access to uevent netlink sockets and prevents containers from
mounting cgroups.

All these restrictions are enforced on the granularity of pid-namespaces
and are enforced only for unprivileged pid-namespaces. The trustme-lsm
offers a securityfs-based userland interface which allows a process from
the pid namespace to drop all privileges which is realized via an
additional data field in the pid_namespace struct. A container is
unprivileged as soon as itself or one of its ancestors has the according
flag set in the pid_namespace struct.

Change-Id: I312dd4af61f4a4d87768fcc1793d97325c2a757c