trustme-lsm: Deny file based access to trustme ICC path
We explicitly deny the access to all files/subdirs in TRUSTME_ICC_PATH.
This prevents an unpriviledged container from deleting or manipulating
the shared filesystem used for ICC. However, a connect to the sock file
inside this path is still possible.

However, we still need to garantee that the TRUSTME_ICC_PATH is the
only path where the shared filesystem is mounted inside a container.

Change-Id: Ie4b448bc84897c4d7a4fab6eddbfaa16eaa49cb9