trustme-lsm: blacklist support
This patch adds support for a static blacklist to the trustme-lsm. A
path is first matched against this blacklist it preceeds the whitelist.
If the path is on the blacklist the access is immediately denied.
Afterwards the path is checked against the whitelist and if it is not
present there, the access is also denied. Therefore, a path must *not*
be on the blacklist *and* be on the whitelist to be allowed.
Furthermore, the patch adds a read-only whitelist which allows to define
paths as read-only accessible for containers.

Change-Id: I6781a4d571b959f3996caa14e39a01b962ad7b2d