trustme-lsm: added hook inode_getattr
It seems that the inode_getattr and path_mknod hooks is also mandatory
for path-based LSMs to catch all edge-cases. Example: touching a file in
a restricted directory still creates the file without these hooks.... We
now have the same path+inode hooks like apparmor and tomoyo. This patch
also adds handling of unnamed sockets/pipes/inodes to the generic
path_decision function which seems to be necessary with using the
inode_getattr hook.
It furthermore employs the file_fcntl and file_ioctl hooks, just to be
sure...

Change-Id: I6372608c7f8ea779d4694de6b9252273ab2cc714