trustme-lsm: mount whitelisting + whitelist
This patch adds a whitelisting functionality for mounts to the
trustme-lsm. It furthermore adds an initial whitelist allowing most of
the necessary mounts with the specific flags etc. The whitelist supports
wildcards for the device, the mount point, the type. It furthermore
supports specifying flags to consider for the comparison and the flags
themselves.

Change-Id: I2f162a7fa5b5198bfec73551b58317c33dfa3aff