capabilities: introduce CAP_SYS_MOUNT
This patch introduces the new capability CAP_SYS_MOUNT. This cap allows
a process to mount/unmount/remount and to create new mount namespaces.
It allows to create mount namespaces with unshare and clone only if
there are NO other namespaces to be created.
It does NOT allow pivot_root or move mounts or the usage of setns
(latter not even for mount namespaces!). For all these scenarios
CAP_SYS_ADMIN is still required.

Change-Id: I5a6ae8dccdb650fccb614e148104bd44063c1792