Deny write access to critical selinuxfs files in case trustme LSM is active
This patch denies access to selinuxfs files which allow to disable or reload
the kernel policy. In case of active trustme-lsm this is necessary as only
the root namespace should be allowed to do those operations. With this patch
containers are enforced to use the loaded policy of the CML init.

Jira -> TM-184

Change-Id: I7cb910f398f02e49e9b55a4252eb00f54cd80fbc