New audio virt via shared audioflinger service [1/2]
This patch introduces sharing of services for multiple containers. It
therefore selectively forwards some of the service requests from
unprivileged containers to a shared context manager normally running in
the privileged management container. It contains a list of services to
be shared that way which currently contains only the two services
necessary for shared access to audio hardware.

Change-Id: I7b8aecdb7af8aab517a325ba03758d85debef0b2